linkedin_insight/requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
python-dotenv
graphviz
Jinja2
lxml

brew install graphviz
